% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_output_nodes.R
\name{make_output_nodes}
\alias{make_output_nodes}
\title{Add Output nodes to a PMML object.}
\usage{
make_output_nodes(name = "OutputField", attributes = NULL,
  expression = NULL, namespace = "4_3")
}
\arguments{
\item{name}{The name of the element to be created}

\item{attributes}{The node attributes to be added}

\item{expression}{Post-processing information to be included in the element.
This expression will be processed by the function_to_pmml function}

\item{namespace}{The namespace of the PMML model}
}
\value{
List of nodes
}
\description{
Add Output nodes to a PMML object.
}
\details{
This function will create a list of nodes with names 'name', attributes 'attributes' and
child elements 'expression'. 'expression' is a string converted to XML similar to th
function_to_pmml function. Meant to create OutputField elements, 'expressions' allows one
to include post-processing transformations to a model. To create multiple such nodes,
all the parameters must be given as lists of equal length.
}
\examples{
# Make two nodes, one with attributes
TwoNodes <- make_output_nodes(
  name = list("OutputField", "OutputField"),
  attributes = list(list(name = "dbl", optype = "continuous"), NULL),
  expression = list("ln(x)", "ln(x/(1-x))")
)
}
\author{
Tridivesh Jena
}
