% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-houseVotes84.R
\docType{data}
\name{houseVotes84}
\alias{houseVotes84}
\title{Modified 1984 United States Congressional Voting Records Database}
\format{A data frame containing: \tabular{ll}{ Class \tab Boolean variable
\cr V1 \tab Boolean variable \cr V2 \tab Boolean variable \cr V3 \tab
Boolean variable \cr V4 \tab Boolean variable \cr V5 \tab Boolean variable
\cr V6 \tab Boolean variable \cr V7 \tab Boolean variable \cr V8 \tab
Boolean variable \cr V9 \tab Boolean variable \cr V10 \tab Boolean variable
\cr V11 \tab Boolean variable \cr V12 \tab Boolean variable \cr V13 \tab
Boolean variable \cr V14 \tab Boolean variable \cr V15 \tab Boolean variable
\cr V16 \tab Boolean variable \cr }}
\description{
This data set includes votes for each of the U.S. House of Representatives
Congressmen on the 16 key votes identified by the CQA.  The CQA lists nine
different types of votes: voted for, paired for, and announced for (these
three simplified to yea), voted against, paired against, and announced
against (these three simplified to nay), voted present, voted present to
avoid conflict of interest, and did not vote or otherwise make a position
known (these three simplified to an unknown disposition).  Originally
containing a binomial variable "class" and 16 other binary variables, those
16 variables have been renamed to simply "V1","V2",...,"V16".
}
\examples{
data(houseVotes84, package = "pmml")
}
\references{
\href{http://archive.ics.uci.edu/ml/datasets/Congressional+Voting+Records}{UCI
Machine Learning Repository}
}
\keyword{datasets}
