% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_values.R
\name{make_values}
\alias{make_values}
\title{Create Values element, most likely to add to a DataDictionary element.}
\usage{
make_values(
  value = NULL,
  displayValue = NULL,
  property = NULL,
  namespace = "4_4"
)
}
\arguments{
\item{value}{The 'value' attribute of each 'Value' element to be created in order.}

\item{displayValue}{The 'displayValue' attribute of each 'Value' element to be created in order.}

\item{property}{The 'property' attribute of each 'Value' element to be created in order.}

\item{namespace}{The namespace of the PMML model}
}
\value{
PMML Values elements.
}
\description{
Create Values element, most likely to add to a DataDictionary element.
}
\details{
This function is used the same way as the \code{make_intervals} function. If certain attributes for an
 element should not be included, they should be input in the list as NULL.
}
\examples{
# define 3 values, none with a 'displayValue' attribute and 1 value
# defined as 'invalid'. The 2nd one is 'valid' by default.
mv <- make_values(
  list(1.1, 2.2, 3.3), list(NULL, NULL, NULL),
  list("valid", NULL, "invalid")
)
}
\seealso{
\code{\link{make_intervals}} to make Interval child elements, \code{\link{add_data_field_children}}
to add these xml fragments to the DataDictionary PMML element.
}
\author{
Tridivesh Jena
}
