% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmml.rules.R
\name{pmml.rules}
\alias{pmml.rules}
\alias{pmml.itemsets}
\title{Generate the PMML representation for a rules or an itemset object from
package \pkg{arules}.}
\usage{
\method{pmml}{rules}(
  model,
  model_name = "arules_Model",
  app_name = "SoftwareAG PMML Generator",
  description = "Association Rules Model",
  copyright = NULL,
  model_version = NULL,
  transforms = NULL,
  ...
)
}
\arguments{
\item{model}{A rules or itemsets object.}

\item{model_name}{A name to be given to the PMML model.}

\item{app_name}{The name of the application that generated the PMML.}

\item{description}{A descriptive text for the Header element of the PMML.}

\item{copyright}{The copyright notice for the model.}

\item{model_version}{A string specifying the model version.}

\item{transforms}{Data transformations.}

\item{...}{Further arguments passed to or from other methods.}
}
\value{
PMML representation of the rules or itemsets object.
}
\description{
Generate the PMML representation for a rules or an itemset object from
package \pkg{arules}.
}
\details{
The model is represented in the PMML AssociationModel format.
}
\references{
\href{https://CRAN.R-project.org/package=arules}{arules: Mining Association
Rules and Frequent Itemsets}
}
\author{
Graham Williams, Michael Hahsler
}
